Struct rugl::clear::Clear

pub struct Clear {
    pub color: Option<[f32; 4]>,
    pub depth: Option<f64>,
    pub stencil: Option<i32>,
}

Clear combines glClearColor, glClearDepth, glClearStencil and glClear into a single procedure, which has the following default usage:

#[macro_use]
extern crate rugl;

fn main() {
    // Use rugl::init() for your code.
    let rugl = rugl::init_headless();

    // Create a clear command that clears the color, depth and stencil.
    let clear_all = rugl!(rugl.clear, {
        color => [0.0, 0.0, 0.0, 1.0],
        depth => 1.0,
        stencil => 0
    });

    // Create a clear command that clears only the color.
    let clear_only_color = rugl!(rugl.clear, {
        color => [0.0, 0.0, 0.0, 1.0]
    });
}

Then run clear_all() or clear_only_color() to run the command. The rugl! macro quickly sets the values and returns a closure that can perform the action. While this is the preferred API, the Clear struct can be used by itself, especially if the clear command needs to be mutated over time.

use rugl;
let rugl = rugl::init_headless();

let mut clear = rugl.clear();
clear.color = Some([0.0, 0.0, 0.0, 1.0]);
// clear.execute();

// The returned function assumes ownership of the Clear struct.
let clear_black = clear.make_execute_fn();
// clear_black();

Fields
color: Option<[f32; 4]>

Sets the clear color

depth: Option<f64>

Sets the clear depth value

stencil: Option<i32>

Sets the clear stencil value

Methods

impl Clear

fn new() -> Clear

Create a new clear object.

fn execute(&self)

Execute the glClear with the set values.

fn make_execute_fn(self) -> Box<Fn()>

Consume the struct and get a closure over execute().